+++ src/schemas/clinical-resolvers/clinicalRegistrationData.ts
@@ -16,16 +16,16 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-import submissionAPI from '../submission/submission-api';
+import submissionAPI from '../../submission/submission-api';
 import {
   convertClinicalRecordToGql,
   convertRegistrationErrorToGql,
   convertClinicalFileErrorToGql,
   convertRegistrationStatsToGql,
   RegistrationErrorData,
-} from './utils';
+} from '../utils';
 import { DeepReadonly } from 'deep-freeze';
-import { ActiveRegistration } from '../submission/submission-entities';
+import { ActiveRegistration } from '../../submission/submission-entities';
 import get from 'lodash/get';
 
 const clinicalRegistrationResolver = {
