+++ src/schemas/resolvers.ts
@@ -1,29 +1,11 @@
-import clinicalRegistrationResolver from './clinicalRegistrationDataResolver';
-import { getClinicalSearchResults, ClinicalSearchQuery } from '../clinical/clinical-service';
-
-// Query Arguments
-type ClinicalVariables = {
-  programShortName: string;
-  filters: ClinicalSearchQuery;
-};
-
-const clinicalResolvers = {
-  Query: {
-    ...clinicalRegistrationResolver,
-    clinicalSearchResults: async (obj: unknown, args: ClinicalVariables) => {
-      const { programShortName, filters } = args;
-
-      const searchResults = (await getClinicalSearchResults(programShortName, filters)) || {
-        searchResults: [],
-      };
-
-      return { ...searchResults, programShortName };
-    },
-  },
-};
+import clinicalRegistrationResolver from './clinical-resolvers/clinicalRegistrationData';
+import clinicalSearchResultResolver from './clinical-resolvers/clinicalSearchResults';
 
 const resolvers = {
-  ...clinicalResolvers,
+  Query: {
+    ...clinicalRegistrationResolver,
+    ...clinicalSearchResultResolver,
+  },
 };
 
 export default resolvers;
